# api/openapi.yaml
openapi: "3.0.3"
info:
  title: TAFE Weather API
  version: "1.0.0"
  description: "RESTful API to read, write, update and delete weather data."
...
paths:
...
  /api/v1/users/{id}:
    get:
      tags:
        - users
      summary: Show a specific user information by user ID
      operationId: showUserById
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/userIdParam"
      responses:
        "200":
          description: OK
        "404":
          $ref: "#/components/responses/404NotFound"
    delete:
      tags:
        - users
      summary: Delete a user by user ID
      operationId: deleteUserById
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/userIdParam"
      responses:
        "204":
          description: The user is deleted successfully.
        "404":
          $ref: "#/components/responses/404NotFound"
...
  /api/v1/account/updatePassword:
    patch:
      tags:
        - account
      summary: User update account password.
      security:
        - bearerAuth: []
      operationId: updateUserPassword
      requestBody:
        $ref: "#/components/requestBodies/updatePasswordInput"
      responses:
        "200":
          description: Password updated successfully.
...
components:
  schemas:
...
    Password:
      type: string
      minLength: 8
      maxLength: 50
      description: User password Input
      x-internal: true
...
    UpdatePassword:
      type: object
      required:
        - currentPassword
        - newPassword
        - confirmNewPassword
      properties:
        currentPassword:
          $ref: "#/components/schemas/Password"
        newPassword:
          $ref: "#/components/schemas/Password"
        confirmNewPassword:
          $ref: "#/components/schemas/Password"
      # x-internal: true
...
    Error:
      type: object
      required:
        - statusCode
        - errors
      properties:
        statusCode:
          type: integer
          description: "error code"
        errors:
          type: array
          items:
            type: object
            required:
              - message
            properties:
              message:
                type: string
              context:
                type: object
                additionalProperties: true
        path:
          type: string
  requestBodies:
...
    updatePasswordInput:
      description: Update user password
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UpdatePassword"
        application/x-www-form-urlencoded:
          schema:
            $ref: "#/components/schemas/UpdatePassword"
...
  parameters:
...
    userIdParam:
      name: id
      in: path
      description: The id of a user to retrive
      required: true
      schema:
        type: string
...
  responses:
...
    404NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT